create column fname and add all students to welcome index
I created a new column in the database for the students first name.
I also added a list of all users onto the welcome page. In the
future the welcome page will be used to check students in and out.
This list will have check in and out buttons to perform this action.